+++ src/Pages/SingUp/SignUp.tsx
@@ -4,22 +4,17 @@ import ProfileSetupForm from '@components/common/ProfileSetupForm';
 import CommonModal from '@components/common/CommonModal';
 import useModal from '@hooks/useModal';
 import SignUpSuccessModal from '@components/Signup/SignUpSuccessModal';
-import { IRegisterRequest, NewUserResponse } from '@/types/AuthTypes';
-import { useSignUpMutation } from '@store/api/authApiSlice';
+import { NewUserResponse } from '@/types/AuthTypes';
 import Spinner from '@components/Auth/Spinner';
-import { useSelector } from 'react-redux';
-import { RootState } from '@store/types';
+import { useSignUp } from '@hooks/useSignup';
 
 function SignUp() {
-  const [signUp, { isLoading: isSingUpLoading }] = useSignUpMutation();
   const navigate = useNavigate();
   const location = useLocation();
   const state = location.state as NewUserResponse;
-  const { authId, authType, nickname, profileImage } = state.data;
+  const { nickname, profileImage } = state.data;
   const { isOpen: isSignUpModalOpen, open: signUpModalOpen } = useModal();
-  const { latitude, longitude, regionCode } = useSelector(
-    (state: RootState) => state.coordinateSetup
-  );
+  const { handleSubmit, isSingUpLoading } = useSignUp(state, signUpModalOpen);
 
   const handleModalOkClick = () => {
     navigate('/', { replace: true });
@@ -34,47 +29,6 @@ function SignUp() {
       />
     );
   }
-
-  const handleSuccessfulSignUp = () => {
-    signUpModalOpen();
-    setTimeout(() => {
-      navigate('/', { replace: true });
-    }, 3000);
-  };
-
-  const handleError = (error: any) => {
-    console.error(error);
-    alert(
-      '회원가입에 실패했습니다. 다시 시도 해주세요. error: ' + error.message
-    );
-    navigate('/auth', { replace: true });
-  };
-
-  const handleSubmit = async (nickname: string) => {
-    const signupData: IRegisterRequest = {
-      authId,
-      authType,
-      nickname,
-      profileImage,
-      addressRequest: {
-        latitude,
-        longitude,
-        regionCode,
-        type: 'main',
-      },
-    };
-
-    try {
-      await signUp(signupData).unwrap();
-      handleSuccessfulSignUp();
-    } catch (error: any) {
-      if (error.status === 409) {
-        alert('중복된 닉네임입니다.');
-      } else {
-        handleError(error);
-      }
-    }
-  };
 
   return (
     <>
